requirements.txt
aiogram==3.20.0
python-dotenv==1.1.0
aiosqlite==0.20.0
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
requests==2.32.3
pyngrok==6.0.1
python-multipart
matplotlib
